refactor(auction_umbrella): separate action.ex in multiple files
