+++ auction_umbrella/apps/auction/lib/auction/fake_repo.ex
@@ -1,26 +1,3 @@
-defmodule Auction.Item do
-  defstruct [:id, :title, :description, :ends_at]
-end
-
-defmodule Auction do
-  alias Auction.{FakeRepo, Item}
-
-  @repo FakeRepo
-
-  def list_items do
-    @repo.all(Item)
-  end
-
-  def get_item(id) do
-    @repo.get!(Item, id)
-  end
-
-  def get_item_by(attrs) do
-    @repo.get_by(Item, attrs)
-  end
-end
-
-
 defmodule Auction.FakeRepo do
   alias Auction.Item
 
@@ -58,5 +35,4 @@ defmodule Auction.FakeRepo do
       end)
     end)
   end
-
 end
