+++ auction_umbrella/apps/auction/lib/auction/item.ex
@@ -1,0 +1,3 @@
+defmodule Auction.Item do
+  defstruct [:id, :title, :description, :ends_at]
+end
